DiscreteRandomMediaGen

Documentation for DiscreteRandomMediaGen.

DiscreteRandomMediaGen.HardSphereModel — Type

A hard sphere (HS) model:

\[\mathrm{e}^{-u(r)}=\left\{ \begin{aligned} 0 &\quad r \le d \\ 1 &\quad r > d \end{aligned} \right.\]

DiscreteRandomMediaGen.SimulationConfig — Type

Configuration for a simulation.

Parameters:

model: the interaction model to be used. Default to HardSphereModel().

volume_fraction: volume fraction of particles. Default to 0.2.

radius_function: a function for generating particle radii. Default to () -> 0.5.

cell_size: size of the peoriodic cell. Default to 40. This leads to 256,000 particles generated in total, which should be enough in most cases.

collisions_per_step: average collisions for each particle before the simulation stops intermediately and checks convergence. Default to 20.

mean_free_time_threshold: threshold for convergence of mean free time. Default to 1e-3.

maximum_steps: maximum number of simulation steps. Default to 100.

save_snapshots: path prefix to save snapshots of the simulation. Default to "", which means no snapshots will be saved. If not empty, snapshots will be named as $(save_snapshots)/snapshot.$step.png for each step.

save_gif: path to save a gif of the simulation. Default to "", which means no gif will be saved. If not empty, a gif will be saved to $(save_fig)/snapshot.gif.

framerate: framerate of the gif. Default to 10.

seed: seed for random number generator. Default to nothing, which means a random seed will be generated.

DiscreteRandomMediaGen.StickyHardSphereModel — Type

A sticky hard sphere (SHS) model specified by Eq. (8.4.37) in Tsang et al. (2001):

\[\mathrm{e}^{-u(r)}=\left\{ \begin{aligned} \frac{d}{12\tau}\delta(r-d) &\quad r \le d \\ 1 &\quad r > d \end{aligned} \right.\]

Parameters:

τ: the stickiness parameter. This has effect only for the SHS model. Default to 1.0.

$\tau \ge \frac{2-\sqrt{2}}{6}$ should hold to ensure that Eq. (8.4.22) in Tsang et al. (2001) has valid solutions.

DiscreteRandomMediaGen.sample_sphere — Function

Sample from the particle configuration described by xml a sphere of radius R centered randomly.

When a particle intersects with the sample sphere, the intersection volume is determined via:

\[\begin{aligned} V & =V\left(R_1, h_1\right)+V\left(R_2, h_2\right) \\ & =\frac{\pi(R+r-d)^2\left(d^2+2 d r-3 r^2+2 d R+6 r R-3 R^2\right)}{12 d} \end{aligned}\]

Then the particle is accepted by probability $\frac{V}{V_{\text{particle}}}$.

Parameters:

xml: the XML Document describing the particle configuration.

R: the radius of the sample sphere.

seed: seed for random number generator. Default to nothing, which means a random seed will be generated.